add scheduledEmail
